+++ LanguageAlfred.WinUI/App.xaml.cs
@@ -1,10 +1,14 @@
 ﻿using CommunityToolkit.Mvvm.DependencyInjection;
 using H.NotifyIcon;
+using LanguageAlfred.WinUI.Services;
+using LanguageAlfred.WinUI.Services.Interfaces;
+using LanguageAlfred.WinUI.ViewModels;
 using LanguageAlfred.WinUI.Views;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.UI;
+using Microsoft.UI.Windowing;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Input;
 using System;
 
 namespace LanguageAlfred.WinUI
@@ -22,7 +26,22 @@ namespace LanguageAlfred.WinUI
 
         protected override void OnLaunched(Microsoft.UI.Xaml.LaunchActivatedEventArgs args)
         {
-            MainWindow = new MainWindow();
+            MainWindow = new Window
+            {
+                Content = new Frame
+                {
+                    Content = new MainView()
+                }
+            };
+
+            IntPtr hWnd = WinRT.Interop.WindowNative.GetWindowHandle(MainWindow);
+            WindowId windowId = Win32Interop.GetWindowIdFromWindow(hWnd);
+            AppWindow appWindow = AppWindow.GetFromWindowId(windowId);
+            var size = new Windows.Graphics.SizeInt32();
+            size.Width = 800;
+            size.Height = 400;
+            appWindow.Resize(size);
+
             MainWindow.Title = "Alfred";
             MainWindow.Activate();
             MainWindow.Hide();
@@ -32,7 +51,7 @@ namespace LanguageAlfred.WinUI
         {
             var services = new ServiceCollection();
             services.AddSingleton<IThemeSelectorService, ThemeSelectorService>();
-            services.AddSingleton<MainWindowViewModel>();
+            services.AddSingleton<MainViewModel>();
             return services.BuildServiceProvider();
         }
     }
