+++ LanguageAlfred.WinUI/LanguageAlfred.WinUI.csproj
@@ -24,6 +24,7 @@
   </PropertyGroup>
   <ItemGroup>
     <None Remove="TrayIconRecources.xaml" />
+    <None Remove="Views\MainView.xaml" />
   </ItemGroup>
 
   <ItemGroup>
@@ -54,6 +55,11 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\LanguageAlfred.VoiceRecognition\LanguageAlfred.VoiceRecognition.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <Page Update="Views\MainView.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
   </ItemGroup>
   <ItemGroup>
     <Page Update="TrayIconRecources.xaml">
