+++ LanguageAlfred.WinUI/Services/Interfaces/IThemeSelectorService.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.UI.Xaml;
+
+namespace LanguageAlfred.WinUI.Services.Interfaces;
+
+public interface IThemeSelectorService
+{
+    public ElementTheme GetTheme();
+    public void SetTheme(ElementTheme theme);
+}
