+++ LanguageAlfred.WinUI/Views/MainView.xaml
@@ -1,13 +1,14 @@
-﻿<Window
-    x:Class="LanguageAlfred.WinUI.MainWindow"
+﻿<Page
+    x:Class="LanguageAlfred.WinUI.Views.MainView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:LanguageAlfred.WinUI"
+    xmlns:local="using:LanguageAlfred.WinUI.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:views="using:LanguageAlfred.WinUI.Views"
     xmlns:tb="using:H.NotifyIcon"
-    mc:Ignorable="d">
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
     <Grid x:Name="AppTitleBar" Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
         <!--<StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Center">
@@ -39,13 +40,12 @@
                             FontWeight="Normal"
                             Text="Theme Mode" />
             <StackPanel x:Name="themePanel" Margin="0,10,0,0">
-                <RadioButton Tag="Default" IsChecked="True" Content="Default" Command="{x:Bind mainWindowViewModel.SetThemeCommand}" CommandParameter="Default"/>
-                <RadioButton Tag="Light" Content="Light"  Command="{x:Bind mainWindowViewModel.SetThemeCommand}" CommandParameter="Light"/>
-                <RadioButton Tag="Dark" Content="Dark"  Command="{x:Bind mainWindowViewModel.SetThemeCommand}" CommandParameter="Dark"/>
+                <RadioButton Tag="Default" IsChecked="True" Content="Default" Command="{x:Bind mainViewModel.SetThemeCommand}" CommandParameter="Default"/>
+                <RadioButton Tag="Light" Content="Light"  Command="{x:Bind mainViewModel.SetThemeCommand}" CommandParameter="Light"/>
+                <RadioButton Tag="Dark" Content="Dark"  Command="{x:Bind mainViewModel.SetThemeCommand}" CommandParameter="Dark"/>
             </StackPanel>
         </StackPanel>
 
         <views:TrayIconView x:Name="TrayIconView" />
     </Grid>
-
-</Window>
+</Page>
